+++ src/game/nlcore/action/baseElement.ts
@@ -29,7 +29,7 @@ export class BaseElement {
         for (let i = 0; i < actions.length; i++) {
             const action = actions[i];
             if (i !== 0) {
-                actions[i - 1]?.contentNode.setInitChild(action.contentNode);
+                actions[i - 1]?.contentNode.setChild(action.contentNode);
             }
         }
         return actions;