+++ src/game/nlcore/action/constructable.ts
@@ -83,9 +83,9 @@ export class Constructable<
         for (let i = 0; i < actions.length; i++) {
             const action = actions[i];
             if (i === 0 && parent) {
-                parent.setInitChild(action.contentNode);
+                parent.setChild(action.contentNode);
             } else if (i > 0) {
-                (actions[i - 1].contentNode)?.setInitChild(action.contentNode);
+                (actions[i - 1].contentNode)?.setChild(action.contentNode);
             }
         }
         return (actions.length) ? actions[0].contentNode : null;