+++ src/game/nlcore/elements/menu.ts
@@ -100,13 +100,13 @@ export class Menu extends Actionable<any, Menu> {
             const node = actions[i].contentNode;
             const child = actions[i + 1]?.contentNode;
             if (child) {
-                node.setInitChild(child);
+                node.setChild(child);
             }
             if (i === this.choices.length - 1 && lastChild) {
-                node.setInitChild(lastChild);
+                node.setChild(lastChild);
             }
             if (i === 0 && parentChild) {
-                parentChild.setInitChild(node);
+                parentChild.setChild(node);
             }
         }
         return actions;